Dinhtrkien

+++ paste-service/src/pasteService.js
@@ -3,7 +3,8 @@ const pasteRepo = require("./pasteRepo");
 
 const cacheService = require("./redis/cacheService");
 
-// Read service URLs from environment variables
+const isCacheEnabled = process.env.ENABLE_REDIS_CACHE === "true";
+
 const SLUG_GENERATOR_URL =
   process.env.SLUG_GENERATOR_URL || "http://localhost:3001";
 const ANALYTICS_SERVICE_URL =
@@ -23,10 +24,11 @@ const pasteService = {
       // Create paste in the database
       const paste = await pasteRepo.createPaste(slug, content, expirationTime);
 
-      // Cache new pastes with a shorter TTL (5 minutes)
-      // Only frequently accessed pastes will stay in cache longer
-      // const INITIAL_CACHE_TTL = 300; // 5 minutes for new pastes
-      // await cacheService.set(slug, paste, INITIAL_CACHE_TTL);
+      // Cache new pastes only if caching is enabled
+      // if (isCacheEnabled) {
+      //   const INITIAL_CACHE_TTL = 300; // 5 minutes for new pastes
+      //   cacheService.set(slug, paste, INITIAL_CACHE_TTL);
+      // }
 
       // // Confirm slug usage to extend cache
       // try {
@@ -149,13 +151,18 @@ const pasteService = {
 
   async getPasteWithoutIncrement(slug) {
     try {
-      // Try to get paste from cache first
-      const cachedPaste = await cacheService.get(slug);
+      let cachedPaste = null;
+      // Try to get paste from cache first if enabled
+      if (isCacheEnabled) {
+        cachedPaste = await cacheService.get(slug);
+      }
 
       if (cachedPaste) {
         // Check if cached paste is expired
         if (pasteRepo.isPasteExpired(cachedPaste)) {
-          await cacheService.delete(slug);
+          if (isCacheEnabled) {
+            await cacheService.delete(slug);
+          }
           return null;
         }
         return cachedPaste;
